pyproject.toml
[tool.poetry]
name = "weather-app"
version = "0.1.0"
description = "Application simple pour afficher la météo d'une ville"
authors = ["Utilisateur <user@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.0"
python-dotenv = "^1.0.0"
colorlog = "^6.8.2"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.8.0"
black = "^24.0.0"
requests-mock = "^1.12.0"

[tool.poetry.scripts]
dev = "ENV=dev python main.py"
prod = "ENV=prod python main.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"